% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/daff-package.r
\docType{package}
\name{daff}
\alias{daff}
\alias{daff-package}
\title{Data diff, patch and merge for R}
\description{
Daff calculates differences between two \code{data.frame}s. This difference can be stored and later used to
patch the original data. Differences can also be made visual by using \code{render_diff} showing what changed.
}
\details{
Storing the difference between data sets allows for tracking or incorporating manual changes to data sets.
Ideally changes to data should be scripted to be reproducable, but there are situations or scenario's where
this is not possible or happens out of your control. \code{daff} can help track these changes.
}

