% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datadiff.R
\name{data_diff}
\alias{data_diff}
\title{data diff}
\usage{
data_diff(data, data_ref)
}
\arguments{
\item{data}{\code{data.frame} to check for changes}

\item{data_ref}{\code{data.frame} reference data frame}
}
\value{
difference object
}
\description{
data diff
}

