% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/diff-data.R
\name{differs_from}
\alias{differs_from}
\title{differs from,}
\usage{
differs_from(data, data_ref, ...)
}
\arguments{
\item{data}{\code{data.frame} to check for changes}

\item{data_ref}{\code{data.frame} reference data frame}

\item{...}{not further specified}
}
\value{
difference object
}
\description{
This is the same function as \code{\link{diff_data}} but with arguments
reversed. This is more useful when using \code{dplyr} and \code{magrittr}
}
\seealso{
diff_data
}

