% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lumberjack.R
\docType{class}
\name{lbj_daff}
\alias{lbj_daff}
\title{Automatically derive daffs with the lumberjack package}
\format{An \code{R6} class object.}
\usage{
lbj_daff
}
\description{
The lbj_daff logger creates a \code{daff} and writes it to \code{daff<n>.csv}
in a directory of choice.
}
\section{Creating a logger}{


\code{filedump$new(dir=file.path(tempdir(),"daff")
   , prefix="step\%03d.csv", verbose=TRUE, ...)}
\tabular{ll}{
  \code{dir}\tab Where to write the daff files.\cr
  \code{filename}\tab filename template, used with \code{\link{sprintf}}
     to create a file name.\cr
  \code{verbose}\tab toggle verbosity\cr
  \code{...}\tab Extra arguments, passed to \code{diff_data} each time.
}
}

\section{Dump options}{


\code{$dump(...)}
\tabular{ll}{
  \code{...}\tab Currently unused.\cr
}
}

\section{Retrieve log data}{


\code{$logdata(simplify=TRUE)} returns a list of data frames, sorted in the
  order returned by \code{base::dir()}
\code{$diffs(simplify=TRUE)} returns a list of \code{daff} objects, sorted
  in the order returned by \code{base::dir()}

\tabular{ll}{
\code{simplify}\tab Simplify lists of length 1 to data.frame?
}
}

\section{Details}{


If \code{dir} does not exist it is created. If
}

\examples{
\dontrun{
 library(lumberjack)
 logger <- lbj_daff$new()
 women \%>>\%
   start_log(logger) \%>>\%
   head() \%>>\%
   {. * 2} \%>>\%
   dump_log()
 L <- logger$logdata()
 L[[2]]
 L <- logger$diffs()
 L[[1]]
}

}
\keyword{datasets}
