% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_diff.R
\name{read_diff}
\alias{read_diff}
\title{Read a diff from a file}
\usage{
read_diff(file)
}
\arguments{
\item{file}{filename or connection}
}
\value{
diff object that can be used in \code{\link{patch_data}}
}
\description{
Read a diff from a file
}

