% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_diff.R
\name{write_diff}
\alias{write_diff}
\title{write a diff to a file}
\usage{
write_diff(diff, file = "diff.csv")
}
\arguments{
\item{diff}{generated with diff_data}

\item{file}{filename or connection}
}
\description{
write a diff to a file
}

