% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datadiff.R
\name{diff_data}
\alias{diff_data}
\title{Do a data diff}
\usage{
diff_data(data_ref, data)
}
\arguments{
\item{data_ref}{\code{data.frame} reference data frame}

\item{data}{\code{data.frame} to check for changes}
}
\value{
difference object
}
\description{
Find differences with a reference data set.
}
\examples{
library(daff)
x <- iris
x[1,1] <- 10
data_diff(x, iris)

}
\seealso{
data_diff
}

