% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patch-data.R
\name{patch_data}
\alias{patch_data}
\title{patch data}
\usage{
patch_data(data, patch)
}
\arguments{
\item{data}{\code{data.frame} that should be patched}

\item{patch}{generated with diff_data}
}
\value{
\code{data.frame} that has been patched.
}
\description{
Patch data with a diff generated by \code{\link{diff_data}}
}
\examples{
library(daff)
x <- iris
#change a value
x[1,1] <- 1000

patch <- diff_data(iris, x)
print(patch)
# apply patch
iris_patched <- patch_data(iris, patch)

iris_patched[1,1] == 1000
}
