% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read-diff.R
\name{write_diff}
\alias{read_diff}
\alias{write_diff}
\title{Write or read a diff to or from a file}
\usage{
write_diff(diff, file = "diff.csv")

read_diff(file)
}
\arguments{
\item{diff}{generated with diff_data}

\item{file}{filename or connection}
}
\value{
diff object that can be used in \code{\link{patch_data}}
}
\description{
Note that type information of the target data.frame is lost when writing a patch to disk.
Using a stored diff to patch a \code{data.frame} will use the column types of the source
\code{data.frame} to determine the target column types. New introduced columns may become \code{characters}.
}

