% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/render-diff.R
\name{render_diff}
\alias{render_diff}
\title{Render a diff table to html}
\usage{
render_diff(diff, file = "")
}
\arguments{
\item{diff}{object}

\item{file}{to write to}
}
\description{
Render a diff table to html
}

